# backend/docs/swagger.yaml
definitions:
  model.FlightCart:
    properties:
      flightID:
        type: integer
      id:
        type: integer
      quantity:
        type: integer
      userID:
        type: integer
    type: object
  model.Promo:
    properties:
      promoBenefit:
        type: integer
      promoCode:
        type: string
      promoDescription:
        type: string
      promoID:
        type: integer
      promoName:
        type: string
      promoPicture:
        type: string
    type: object
  model.Question:
    properties:
      question:
        type: string
      questionID:
        type: integer
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger TPA JZ23-2
  version: "1.0"
paths:
  /api/admin/delete-promo:
    delete:
      description: Delete a promo by PromoID
      parameters:
      - description: Promo ID
        in: query
        name: PromoID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Promo Deleted
          schema:
            type: string
      summary: Delete Promo
  /api/admin/update-promo:
    put:
      consumes:
      - application/json
      description: Update Existing Promo
      parameters:
      - description: Update Promo
        in: body
        name: promo
        required: true
        schema:
          $ref: '#/definitions/model.Promo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Promo'
      summary: Update Existing Promo
  /api/cart/add-flight-cart:
    post:
      consumes:
      - application/json
      description: Create a new Flight Cart
      parameters:
      - description: Flight Cart details
        in: body
        name: promo
        required: true
        schema:
          $ref: '#/definitions/model.FlightCart'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.FlightCart'
      summary: Create Flight Cart
  /api/user/get-user:
    get:
      consumes:
      - application/json
      description: get users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Question'
            type: array
      summary: List Users
      tags:
      - user
swagger: "2.0"